Updated refinery upgrade panel

+++ Assets/Scripts/RefineryController.cs
@@ -153,7 +153,7 @@ public class RefineryController : MonoBehaviour, IDataPersistence
 
         // If there is a lobby ad display added to ad delegator, try to show the lobby ad reward
         if (adDelegator.lobbyAdDisplay) {
-            StartCoroutine(adDelegator.TryShowLobbyReward((long) cashMadeThisMine));
+            StartCoroutine(adDelegator.TryShowLobbyReward(cashMadeThisMine));
         }
 
         playerState.UpdateHighestMined(cashMadeThisMine);
